% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{add_abj_logo}
\alias{add_abj_logo}
\title{Add ABJ logo to image}
\usage{
add_abj_logo(img, logo = system.file("logo.png", package = "logo"),
  prop = 0.1, position = "bottom")
}
\arguments{
\item{img}{image path}

\item{logo}{logo path. Defalts to system file}

\item{prop}{size as proportion of image. Defaults to 10\%}

\item{position}{top or bottom}
}
\value{
invisible image
}
\description{
Add ABJ logo to image
}
